refactor: CoinDetail coin 전달 로직 수정
