+++ src/components/CoinDetail/Chart/CoinDetailChart.style.ts
@@ -1,9 +1,5 @@
 import styled from '@emotion/styled';
 
-export const Container = styled.div`
-  height: 100%;
-`;
-
 export const ChartWrapper = styled.div`
   position: relative;
 `;
@@ -12,10 +8,11 @@ export const Select = styled.select`
   position: absolute;
   z-index: 9999;
   top: 0;
-  left: 90px;
+  right: 140px;
   padding: 5px;
   border-radius: 4px;
   border: solid;
+  font-size: 10px;
   ${({theme: {colors}}) => {
     return {
       backgroundColor: colors.BACKGROUND_SUB,
