+++ src/components/CoinDetail/Chart/CoinDetailChartOptions.tsx
