+++ src/components/CoinDetail/CoinDetail.tsx
@@ -1,28 +1,13 @@
-import {useCoinDetail} from '@/hooks';
-import Chart from './Chart';
-import * as S from './CoinDetail.style';
-import CoinDetailChartOptions from './CoinDetailChartOptions';
-import CoinDetailInfo from './CoinDetailInfo';
+import CoinDetailChart from './Chart/CoinDetailChart';
+import CoinDetailInfo from './Info/CoinDetailInfo';
 
-function CoinDetail({coin}: CoinDetailProp) {
-  const {
-    chartData,
-    convertedCoin,
-    candleState: {type, unit},
-    onChange,
-  } = useCoinDetail(coin);
-
+const CoinDetail = () => {
   return (
-    <S.Container>
-      <CoinDetailInfo coin={convertedCoin} />
-      <S.ChartWrapper>
-        <S.Select onChange={onChange} defaultValue={`${type} ${unit}`}>
-          <CoinDetailChartOptions />
-        </S.Select>
-        <Chart type={'candlestick'} {...chartData} />
-      </S.ChartWrapper>
-    </S.Container>
+    <>
+      <CoinDetailInfo />
+      <CoinDetailChart />
+    </>
   );
-}
+};
 
 export default CoinDetail;
