+++ src/components/CoinDetail/Info/CoinDetailInfo.style.ts
