+++ src/hooks/queries/useDayCandleQuery.ts
@@ -6,6 +6,7 @@ const useDayCandleQuery = (request: DayCandleRequest) => {
   return useQuery([QUERY_KEYS.candles, 'day', request.queries.market], {
     queryFn: async () => await getDaysCandles(request),
     staleTime: 1000 * 60 * 60 * 24,
+    enabled: !!request.queries.market,
   });
 };
