+++ src/hooks/queries/useMonthCandleQuery.ts
@@ -6,6 +6,7 @@ const useMonthCandleQuery = (request: MonthCandleRequest) => {
   return useQuery([QUERY_KEYS.candles, 'month', request.queries.market], {
     queryFn: async () => await getMonthsCandles(request),
     staleTime: 1000 * 60 * 60 * 24 * 28,
+    enabled: !!request.queries.market,
   });
 };
