+++ src/hooks/queries/useWeekCandleQuery.ts
@@ -6,6 +6,7 @@ const useWeekCandleQuery = (request: WeekCandleRequest) => {
   return useQuery([QUERY_KEYS.candles, 'week', request.queries.market], {
     queryFn: async () => await getWeeksCandles(request),
     staleTime: 1000 * 60 * 60 * 24 * 7,
+    enabled: !!request.queries.market,
   });
 };
